Fix bug, usability tweak

+++ digitanks/src/game/digitanks/digitanksteam.h
@@ -57,7 +57,7 @@ public:
 	void						CountFleetPoints();
 	size_t						GetTotalFleetPoints() { return m_iTotalFleetPoints; };
 	size_t						GetUsedFleetPoints() { return m_iUsedFleetPoints; };
-	size_t						GetUnusedFleetPoints() { return GetTotalFleetPoints() - GetUsedFleetPoints(); };
+	size_t						GetUnusedFleetPoints() { return (int)GetTotalFleetPoints() - (int)GetUsedFleetPoints(); };
 
 	void						CountScore();
 	size_t						GetScore() { return m_iScore; };
